+++ src/paths.rs
@@ -4,7 +4,7 @@ use crate::options::CLIOptions;
 
 pub struct Paths {
 	pub cache_dir_path: PathBuf,
-	pub config_file_path: Option<PathBuf>,
+	pub config_file_path: PathBuf,
 	pub data_dir_path: PathBuf,
 	pub db_file_path: PathBuf,
 	pub log_file_path: Option<PathBuf>,
@@ -21,7 +21,7 @@ impl Default for Paths {
 	fn default() -> Self {
 		Self {
 			cache_dir_path: ["."].iter().collect(),
-			config_file_path: None,
+			config_file_path: [".", "polaris.toml"].iter().collect(),
 			data_dir_path: ["."].iter().collect(),
 			db_file_path: [".", "db.sqlite"].iter().collect(),
 			log_file_path: Some([".", "polaris.log"].iter().collect()),
@@ -40,7 +40,7 @@ impl Default for Paths {
 			local_app_data.join(["Permafrost", "Polaris"].iter().collect::<PathBuf>());
 		Self {
 			cache_dir_path: install_directory.clone(),
-			config_file_path: None,
+			config_file_path: install_directory.join("polaris.toml"),
 			data_dir_path: install_directory.clone(),
 			db_file_path: install_directory.join("db.sqlite"),
 			log_file_path: Some(install_directory.join("polaris.log")),
@@ -58,10 +58,12 @@ impl Paths {
 				.map(PathBuf::from)
 				.map(|p| p.join("db.sqlite"))
 				.unwrap_or(defaults.db_file_path),
-			config_file_path: None,
 			cache_dir_path: option_env!("POLARIS_CACHE_DIR")
 				.map(PathBuf::from)
 				.unwrap_or(defaults.cache_dir_path),
+			config_file_path: option_env!("POLARIS_CONFIG_DIR")
+				.map(|p| [p, "polaris.toml"].iter().collect())
+				.unwrap_or(defaults.config_file_path),
 			data_dir_path: option_env!("POLARIS_DATA_DIR")
 				.map(PathBuf::from)
 				.unwrap_or(defaults.data_dir_path),
@@ -89,7 +91,7 @@ impl Paths {
 			path.clone_into(&mut paths.cache_dir_path);
 		}
 		if let Some(path) = &cli_options.config_file_path {
-			paths.config_file_path = Some(path.clone());
+			path.clone_into(&mut paths.config_file_path);
 		}
 		if let Some(path) = &cli_options.data_dir_path {
 			path.clone_into(&mut paths.data_dir_path);