+++ src/server/axum.rs
@@ -59,12 +59,6 @@ impl FromRef<App> for app::config::Manager {
 	}
 }
 
-impl FromRef<App> for app::ddns::Manager {
-	fn from_ref(app: &App) -> Self {
-		app.ddns_manager.clone()
-	}
-}
-
 impl FromRef<App> for app::peaks::Manager {
 	fn from_ref(app: &App) -> Self {
 		app.peaks_manager.clone()
@@ -77,26 +71,8 @@ impl FromRef<App> for app::playlist::Manager {
 	}
 }
 
-impl FromRef<App> for app::user::Manager {
-	fn from_ref(app: &App) -> Self {
-		app.user_manager.clone()
-	}
-}
-
-impl FromRef<App> for app::settings::Manager {
-	fn from_ref(app: &App) -> Self {
-		app.settings_manager.clone()
-	}
-}
-
 impl FromRef<App> for app::thumbnail::Manager {
 	fn from_ref(app: &App) -> Self {
 		app.thumbnail_manager.clone()
-	}
-}
-
-impl FromRef<App> for app::vfs::Manager {
-	fn from_ref(app: &App) -> Self {
-		app.vfs_manager.clone()
 	}
 }