+++ src/server/axum/test.rs
@@ -23,7 +23,7 @@ impl TestService for AxumTestService {
 
 		let paths = Paths {
 			cache_dir_path: ["test-output", test_name].iter().collect(),
-			config_file_path: None,
+			config_file_path: output_dir.join("polaris.toml"),
 			data_dir_path: ["test-output", test_name].iter().collect(),
 			db_file_path: output_dir.join("db.sqlite"),
 			#[cfg(unix)]