+++ CHANGELOG.md
@@ -18,7 +18,8 @@
 - The `/thumbnail` endpoint supports a new size labeled `tiny` which returns 40x40px images.
 - Persistent data, such as playlists, is now saved in a directory that may be configured with the `--data` CLI option or the `POLARIS_DATA_DIR` environment variable.
 - Removed last.fm integration due to maintenance concerns (abandoned libraries, broken account linking) and mismatch with project goals.
-- Removed `/config` API endpoint.
+- Removed the `/config` API endpoint.
+- Removed the `/ddns` API endpoints, merged into the existing `/settings` endpoints.
 
 ### Web client